+++ main/RDRanching/Recipies/Breeding/AgusCowBreed.cs
@@ -1,23 +1,17 @@
 ﻿namespace RD.Framework.main.RDRanching.Recipies.Breeding
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
-    using System.ComponentModel;
     using Eco.Core.Items;
     using Eco.Gameplay.Components;
     using Eco.Gameplay.Items.Recipes;
     using Eco.Gameplay.Skills;
     using Eco.Shared.Localization;
-    using RD.Framework.main.RDRanching.Items;
     using RD.Framework.main.RDRanching.Items.Animals.AdultAnimals;
     using RD.Framework.main.RDRanching.Items.obj;
+    using RD.Framework.main.RDRanching.Items;
     using RD.Framework.main.RDSkills.Ranching;
 
     [RequiresSkill(typeof(RancherSkill), 1)]
-    [Ecopedia("Items", "Products", subPageName: "Angus Cow")]
+    [Ecopedia("Items", "Products", subPageName: "Angu Cow")]
     public partial class AngusCowRecipe : RecipeFamily
     {
         public AngusCowRecipe()
@@ -36,7 +30,7 @@
                     new CraftingElement<AngusCowItem>(1)
                 });
             Recipes = new List<Recipe> { recipe };
-            ExperienceOnCraft = 5;
+            ExperienceOnCraft = 5f;
 
             LaborInCalories = CreateLaborInCaloriesValue(70, typeof(RancherSkill));
 
@@ -47,40 +41,6 @@
             ModsPostInitialize();
 
             CraftingComponent.AddRecipe(tableType: typeof(NurseryObject), recipe: this);
-        }
-        partial void ModsPreInitialize();
-        partial void ModsPostInitialize();
-    }
-
-    public partial class DairyCowRecipe : RecipeFamily
-    {
-        public DairyCowRecipe()
-        {
-            var recipe = new Recipe();
-            recipe.Init(
-                name: "AngusCow",
-                displayName: Localizer.DoStr("Angus Cow"),
-                ingredients: new List<IngredientElement>
-                {
-                    new IngredientElement(typeof(CowItem), 2, true),
-                    new IngredientElement(typeof(AnimalFeedItem), 6, true),
-                },
-                items: new List<CraftingElement>
-                {
-                    new CraftingElement<DairyCowItem>(1)
-                });
-            Recipes = new List<Recipe> { recipe };
-            ExperienceOnCraft = 5;
-
-            LaborInCalories = CreateLaborInCaloriesValue(70, typeof(RancherSkill));
-
-            CraftMinutes = CreateCraftTimeValue(beneficiary: typeof(DairyCowRecipe), start: 5.0f, skillType: typeof(RancherSkill));
-
-            ModsPreInitialize();
-            Initialize(displayText: Localizer.DoStr("Angus Cow"), recipeType: typeof(DairyCowRecipe));
-            ModsPostInitialize();
-
-            CraftingComponent.AddRecipe(tableType: typeof(BarnObject), recipe: this);
         }
         partial void ModsPreInitialize();
         partial void ModsPostInitialize();
